requirements.txt
streamlit
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-aiplatform
google-auth-httplib2
google-generativeai
requests